+++ src/Services/CRM/CRMServiceBuilder.php
@@ -719,4 +719,16 @@ class CRMServiceBuilder extends AbstractServiceBuilder
 
         return $this->serviceCache[__METHOD__];
     }
+
+    public function userfieldConfig(): Userfieldconfig\Service\Userfieldconfig
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Userfieldconfig\Service\Userfieldconfig(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
 }